Modify changeSelectedMark function
The changeSelectedMark function was modified because the use of e.target
caused unexpected behavior with the event listener, and will cause a
malfunction in the button click. this malfunction happen when the image
inside the button is clicked and does not have the id attribute that is
used in the logic defined to change classes and image sources.